requirements.txt
pyttsx3
SpeechRecognition
wikipedia-api
requests
pyjokes
psutil
comtypes
pyaudio
